requirement.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-core
langchain-llms
langchain-embeddings
streamlit
requests
pydantic
